Move layer_manger to layer/manager.py.
