+++ microproxy/layer/manager.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import
+
 from tornado import gen
 from tornado import iostream
